Joe says: > 	* Linker.cc: Added (unsigned int) cast to get rid of gcc-2.6.0 warning: > 	Linker.cc:374: warning: unsigned int format, pointer arg (arg 7) This won't work on a 64-bit-address machine with 32-bit ints.  Pointers should be printed by using casts to unsigned long, which should always be wide enough not to lose any precision. (We don't support such machines now, but you might as well fix things now that will make an Alpha port harder).

+++ ptclassic/src/kernel/Linker.cc
@@ -372,7 +372,7 @@ int Linker::multiLink (int argc, char** argv) {
 	// DYNLIB is hppa specific
 	sprintf (command, "%s %s -A %s %s %x %s -o ",
 		 LOADER, LOADOPTS, symTableName, LOC_OPT,
-		 (unsigned int)availMem, DYNLIB);
+		 (unsigned long)availMem, DYNLIB);
 	D( Error::message(command); )
 	StringList cmd = command;
 	cmd << tname;
